[CLN] payment_razorpay_oauth: merge into `payment_razorpay`
This commit merges the Razorpay OAuth flow into the main `payment_razorpay`
module, integrating both flows into a single one, and removing the
`payment_razorpay_oauth` module.
