+++ addons/payment_razorpay/models/payment_transaction.py
@@ -41,7 +41,7 @@ class PaymentTransaction(models.Model):
         order_id = self._razorpay_create_order(customer_id)['id']
         return {
             'razorpay_key_id': self.provider_id.razorpay_key_id,
-            'razorpay_public_token': self.provider_id._razorpay_get_public_token(),
+            'razorpay_public_token': self.provider_id.razorpay_public_token,
             'razorpay_customer_id': customer_id,
             'is_tokenize_request': self.tokenize,
             'razorpay_order_id': order_id,
