--- addons/payment_razorpay_oauth/const.py
@@ -1,3 +1,0 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-OAUTH_URL = 'https://razorpay.api.odoo.com/api/razorpay/1'
